Mod 3 Changes
Refactored code from Mod 2. The sign in form is a non-model-backed
form, and I had been using bootstrap_form_for to render the form.
Refactored the code to use form_tag, which simplified the code, and
allowed user input to be saved and sent back to the form in the event
of invalid user input.